ocr_pipeline/requirements.txt
# Core PDF processing
PyMuPDF==1.23.22
pymupdf4llm==0.0.10

# OCR engines for scanned documents
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.9.0.80

# Advanced OCR and layout analysis
easyocr==1.7.1
paddlepaddle==2.6.1
paddleocr==2.7.3

# Text processing and structure preservation
python-docx==1.1.0
pandas==2.2.2
numpy==1.26.4

# Image processing and enhancement
scikit-image==0.22.0
matplotlib==3.8.4

# Configuration management
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Logging and progress tracking
tqdm==4.66.4
rich==13.7.1
loguru==0.7.2

# Testing
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development tools
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# Optional: AI-powered text extraction
# transformers==4.40.1
# torch==2.3.0